+++ src/streaming_kzg/space.rs
@@ -284,71 +284,3 @@ where
         }
     }
 }
-
-#[test]
-fn test_open_multi_points() {
-    use crate::ark_std::UniformRand;
-    use crate::streaming_kzg::evaluate_be;
-    use ark_bls12_381::{Bls12_381, Fr};
-    use ark_ff::Field;
-    use ark_poly::univariate::DensePolynomial;
-    use ark_poly::UVPolynomial;
-    use ark_std::test_rng;
-
-    let max_msm_buffer = 1 << 20;
-    let rng = &mut test_rng();
-    // f = 80*x^6 + 80*x^5 + 88*x^4 + 3*x^3 + 73*x^2 + 7*x + 24
-    let polynomial = [
-        Fr::from(80u64),
-        Fr::from(80u64),
-        Fr::from(88u64),
-        Fr::from(3u64),
-        Fr::from(73u64),
-        Fr::from(7u64),
-        Fr::from(24u64),
-    ];
-    let polynomial_stream = &polynomial[..];
-    let beta = Fr::from(53u64);
-
-    let time_ck = CommitterKey::<Bls12_381>::new(200, 3, rng);
-    let space_ck = CommitterKeyStream::from(&time_ck);
-
-    let (remainder, _commitment) = space_ck.open_multi_points(
-        &polynomial_stream,
-        &[beta.square(), beta, -beta],
-        max_msm_buffer,
-    );
-    let evaluation_remainder = evaluate_be(&remainder, &beta);
-    assert_eq!(evaluation_remainder, Fr::from(1807299544171u64));
-
-    let (remainder, _commitment) =
-        space_ck.open_multi_points(&polynomial_stream, &[beta], max_msm_buffer);
-    assert_eq!(remainder.len(), 1);
-
-    // get a random polynomial with random coefficient,
-    let polynomial = DensePolynomial::rand(100, rng).coeffs().to_vec();
-    let polynomial_stream = &polynomial[..];
-    let beta = Fr::rand(rng);
-    let (_, evaluation_proof_batch) =
-        space_ck.open_multi_points(&polynomial_stream, &[beta], max_msm_buffer);
-    let (_, evaluation_proof_single) = space_ck.open(&polynomial_stream, &beta, max_msm_buffer);
-    assert_eq!(evaluation_proof_batch, evaluation_proof_single);
-
-    let (remainder, _evaluation_poof) = space_ck.open_multi_points(
-        &polynomial_stream,
-        &[beta, -beta, beta.square()],
-        max_msm_buffer,
-    );
-    let expected_evaluation = evaluate_be(&remainder, &beta);
-    let obtained_evaluation = evaluate_be(&polynomial, &beta);
-    assert_eq!(expected_evaluation, obtained_evaluation);
-    let expected_evaluation = evaluate_be(&remainder, &beta.square());
-    let obtained_evaluation = evaluate_be(&polynomial, &beta.square());
-    assert_eq!(expected_evaluation, obtained_evaluation);
-    // let expected_evaluation = evaluate_be(&remainder, &beta.square());
-    // let obtained_evaluation = evaluate_be(&polynomial, &beta.square());
-    // assert_eq!(expected_evaluation, obtained_evaluation);
-    // let expected_evaluation = evaluate_be(&remainder, &beta.square());
-    // let obtained_evaluation = evaluate_be(&polynomial, &beta.square());
-    // assert_eq!(expected_evaluation, obtained_evaluation);
-}
