+++ src/streaming_kzg/tests.rs
@@ -2,12 +2,38 @@ use ark_bls12_381::{Bls12_381, Fr};
 use ark_poly::univariate::DensePolynomial;
 use ark_poly::UVPolynomial;
 use ark_std::vec::Vec;
-use ark_std::UniformRand;
+use ark_std::{UniformRand, Zero};
 
 use crate::streaming_kzg::space::CommitterKeyStream;
 use crate::streaming_kzg::time::CommitterKey;
-use crate::streaming_kzg::{evaluate_le, VerifierKey};
+use crate::streaming_kzg::{vanishing_polynomial, VerifierKey};
+use ark_ff::Field;
+use ark_std::borrow::Borrow;
 use ark_std::iterable::{Iterable, Reverse};
+
+/// Polynomial evaluation, assuming that the
+/// coefficients are in little-endian.
+#[inline]
+fn evaluate_le<F>(polynomial: &[F], x: &F) -> F
+where
+    F: Field,
+{
+    evaluate_be(polynomial.iter().rev(), x)
+}
+
+/// Polynomial evaluation, assuming that the
+/// coeffients are in big-endian.
+#[inline]
+fn evaluate_be<I, F>(polynomial: I, x: &F) -> F
+where
+    F: Field,
+    I: IntoIterator,
+    I::Item: Borrow<F>,
+{
+    polynomial
+        .into_iter()
+        .fold(F::zero(), |previous, c| previous * x + c.borrow())
+}
 
 #[test]
 fn test_commitment_consistency() {
@@ -26,7 +52,7 @@ fn test_commitment_consistency() {
 }
 
 #[test]
-fn test_srs() {
+fn test_ck_consistency() {
     use ark_bls12_381::Bls12_381;
 
     let rng = &mut ark_std::test_rng();
@@ -96,4 +122,136 @@ fn test_open_multipoints_correctness() {
     );
 
     assert!(verification_result.is_ok());
+}
+
+#[test]
+fn test_vanishing_polynomial() {
+    use ark_bls12_381::Fr as F;
+    use ark_ff::Zero;
+
+    let points = [F::from(10u64), F::from(5u64), F::from(13u64)];
+    let zeros = vanishing_polynomial(&points);
+    assert_eq!(evaluate_le(&zeros, &points[0]), F::zero());
+    assert_eq!(evaluate_le(&zeros, &points[1]), F::zero());
+    assert_eq!(evaluate_le(&zeros, &points[2]), F::zero());
+}
+
+#[test]
+fn test_srs() {
+    use ark_bls12_381::Bls12_381;
+
+    let rng = &mut ark_std::test_rng();
+    let ck = CommitterKey::<Bls12_381>::new(10, 3, rng);
+    let vk = VerifierKey::from(&ck);
+    // Make sure that there are enough elements for the entire array.
+    assert_eq!(ck.powers_of_g.len(), 11);
+    assert_eq!(ck.powers_of_g2, &vk.powers_of_g2[..]);
+}
+
+#[test]
+fn test_trivial_commitment() {
+    use ark_bls12_381::Bls12_381;
+    use ark_bls12_381::Fr;
+    use ark_poly::univariate::DensePolynomial;
+    use ark_poly::UVPolynomial;
+    use ark_std::One;
+
+    let rng = &mut ark_std::test_rng();
+    let ck = CommitterKey::<Bls12_381>::new(10, 3, rng);
+    let vk = VerifierKey::from(&ck);
+    let polynomial = DensePolynomial::from_coefficients_slice(&[Fr::zero(), Fr::one(), Fr::one()]);
+    let alpha = Fr::zero();
+
+    let commitment = ck.commit(&polynomial);
+    let (evaluation, proof) = ck.open(&polynomial, &alpha);
+    assert_eq!(evaluation, Fr::zero());
+    assert!(vk.verify(&commitment, &alpha, &evaluation, &proof).is_ok())
+}
+
+#[test]
+fn test_commitment() {
+    use ark_bls12_381::Bls12_381;
+    use ark_bls12_381::Fr;
+    use ark_poly::univariate::DensePolynomial;
+    use ark_poly::Polynomial;
+    use ark_poly::UVPolynomial;
+
+    let rng = &mut ark_std::test_rng();
+    let ck = CommitterKey::<Bls12_381>::new(100, 3, rng);
+    let vk = VerifierKey::from(&ck);
+    let polynomial = DensePolynomial::rand(100, rng);
+    let alpha = Fr::zero();
+
+    let commitment = ck.commit(&polynomial);
+    let (evaluation, proof) = ck.open(&polynomial, &alpha);
+    let expected_evaluation = polynomial.evaluate(&alpha);
+    assert_eq!(evaluation, expected_evaluation);
+    assert!(vk.verify(&commitment, &alpha, &evaluation, &proof).is_ok())
+}
+
+#[test]
+fn test_open_multi_points() {
+    use crate::ark_std::UniformRand;
+    use ark_bls12_381::{Bls12_381, Fr};
+    use ark_ff::Field;
+    use ark_poly::univariate::DensePolynomial;
+    use ark_poly::UVPolynomial;
+    use ark_std::test_rng;
+
+    let max_msm_buffer = 1 << 20;
+    let rng = &mut test_rng();
+    // f = 80*x^6 + 80*x^5 + 88*x^4 + 3*x^3 + 73*x^2 + 7*x + 24
+    let polynomial = [
+        Fr::from(80u64),
+        Fr::from(80u64),
+        Fr::from(88u64),
+        Fr::from(3u64),
+        Fr::from(73u64),
+        Fr::from(7u64),
+        Fr::from(24u64),
+    ];
+    let polynomial_stream = &polynomial[..];
+    let beta = Fr::from(53u64);
+
+    let time_ck = CommitterKey::<Bls12_381>::new(200, 3, rng);
+    let space_ck = CommitterKeyStream::from(&time_ck);
+
+    let (remainder, _commitment) = space_ck.open_multi_points(
+        &polynomial_stream,
+        &[beta.square(), beta, -beta],
+        max_msm_buffer,
+    );
+    let evaluation_remainder = evaluate_be(&remainder, &beta);
+    assert_eq!(evaluation_remainder, Fr::from(1807299544171u64));
+
+    let (remainder, _commitment) =
+        space_ck.open_multi_points(&polynomial_stream, &[beta], max_msm_buffer);
+    assert_eq!(remainder.len(), 1);
+
+    // get a random polynomial with random coefficient,
+    let polynomial = DensePolynomial::rand(100, rng).coeffs().to_vec();
+    let polynomial_stream = &polynomial[..];
+    let beta = Fr::rand(rng);
+    let (_, evaluation_proof_batch) =
+        space_ck.open_multi_points(&polynomial_stream, &[beta], max_msm_buffer);
+    let (_, evaluation_proof_single) = space_ck.open(&polynomial_stream, &beta, max_msm_buffer);
+    assert_eq!(evaluation_proof_batch, evaluation_proof_single);
+
+    let (remainder, _evaluation_poof) = space_ck.open_multi_points(
+        &polynomial_stream,
+        &[beta, -beta, beta.square()],
+        max_msm_buffer,
+    );
+    let expected_evaluation = evaluate_be(&remainder, &beta);
+    let obtained_evaluation = evaluate_be(&polynomial, &beta);
+    assert_eq!(expected_evaluation, obtained_evaluation);
+    let expected_evaluation = evaluate_be(&remainder, &beta.square());
+    let obtained_evaluation = evaluate_be(&polynomial, &beta.square());
+    assert_eq!(expected_evaluation, obtained_evaluation);
+    // let expected_evaluation = evaluate_be(&remainder, &beta.square());
+    // let obtained_evaluation = evaluate_be(&polynomial, &beta.square());
+    // assert_eq!(expected_evaluation, obtained_evaluation);
+    // let expected_evaluation = evaluate_be(&remainder, &beta.square());
+    // let obtained_evaluation = evaluate_be(&polynomial, &beta.square());
+    // assert_eq!(expected_evaluation, obtained_evaluation);
 }
