+++ src/streaming_kzg/time.rs
@@ -160,56 +160,3 @@ impl<E: PairingEngine> CommitterKey<E> {
         self.open_multi_points(&batched_polynomial, eval_points)
     }
 }
-
-#[test]
-fn test_srs() {
-    use ark_bls12_381::Bls12_381;
-
-    let rng = &mut ark_std::test_rng();
-    let ck = CommitterKey::<Bls12_381>::new(10, 3, rng);
-    let vk = VerifierKey::from(&ck);
-    // Make sure that there are enough elements for the entire array.
-    assert_eq!(ck.powers_of_g.len(), 11);
-    assert_eq!(ck.powers_of_g2, &vk.powers_of_g2[..]);
-}
-
-#[test]
-fn test_trivial_commitment() {
-    use ark_bls12_381::Bls12_381;
-    use ark_bls12_381::Fr;
-    use ark_poly::univariate::DensePolynomial;
-    use ark_poly::UVPolynomial;
-    use ark_std::One;
-
-    let rng = &mut ark_std::test_rng();
-    let ck = CommitterKey::<Bls12_381>::new(10, 3, rng);
-    let vk = VerifierKey::from(&ck);
-    let polynomial = DensePolynomial::from_coefficients_slice(&[Fr::zero(), Fr::one(), Fr::one()]);
-    let alpha = Fr::zero();
-
-    let commitment = ck.commit(&polynomial);
-    let (evaluation, proof) = ck.open(&polynomial, &alpha);
-    assert_eq!(evaluation, Fr::zero());
-    assert!(vk.verify(&commitment, &alpha, &evaluation, &proof).is_ok())
-}
-
-#[test]
-fn test_commitment() {
-    use ark_bls12_381::Bls12_381;
-    use ark_bls12_381::Fr;
-    use ark_poly::univariate::DensePolynomial;
-    use ark_poly::Polynomial;
-    use ark_poly::UVPolynomial;
-
-    let rng = &mut ark_std::test_rng();
-    let ck = CommitterKey::<Bls12_381>::new(100, 3, rng);
-    let vk = VerifierKey::from(&ck);
-    let polynomial = DensePolynomial::rand(100, rng);
-    let alpha = Fr::zero();
-
-    let commitment = ck.commit(&polynomial);
-    let (evaluation, proof) = ck.open(&polynomial, &alpha);
-    let expected_evaluation = polynomial.evaluate(&alpha);
-    assert_eq!(evaluation, expected_evaluation);
-    assert!(vk.verify(&commitment, &alpha, &evaluation, &proof).is_ok())
-}
